add chart locals helper
